Gameloop Balanced

+++ Scripts/Weapons/Bullet.cs
@@ -4,7 +4,7 @@ public partial class Bullet : Node2D
 {
     private float _speed = 300f;
     private float _lifeSeconds = 2f;
-    public static float Damage = 50f;
+    public static int Damage = 50;
     private Vector2 _direction = Vector2.Right;
 
     public override void _Ready()
